Add EntryValidator to allow automatic fallback to default value
If you provide an EntryValidator for an ConfigEntry and it returns `false` in on of the `#getAs...()` methods, it will automatically use the default value but without changing the current/normal/user value. This ensurs, that the file is not modified.

Additionally, the user is informed about a value being invalid and the default value as an example.